create product schema

+++ app/utils/unlink-file.js
@@ -1,9 +1,10 @@
 const path = require("path");
 const fs = require("fs");
 const deleteFileInPublic = (file) => {
-  const filePath = path.join(__dirname, "..", "..", "public", file);
-  console.log(filePath);
-  fs.unlinkSync(filePath);
+  if (file) {
+    const filePath = path.join(__dirname, "..", "..", "public", file);
+    if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
+  }
 };
 
 module.exports = {
